backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
tensorflow-cpu==2.15.0
keras==2.15.0
yfinance==0.2.33
joblib==1.3.2
matplotlib==3.8.2
supabase==2.3.0
python-dotenv==1.0.0
